refactor(fbs-common): verplaats Profiel-laag uit berichtenmagazijn
- ProfielServiceClient + DTO's, ToestemmingGeweigerd-exception + mapper,
  EndpointValidator + test naar fbs-common.profiel
- Nieuwe ProfielServiceFoutException + mapper (503 + Retry-After)
- Magazijn-imports gepatcht; gedragsneutraal voor #541-flow.

Bereidt voor: gedeelde Profiel-laag voor MagazijnResolver in sessiecache.
