+++ libraries/fbs-common/src/main/kotlin/nl/rijksoverheid/moz/fbs/common/profiel/ProfielServiceClient.kt
@@ -1,4 +1,4 @@
-package nl.rijksoverheid.moz.fbs.berichtenmagazijn.validatie
+package nl.rijksoverheid.moz.fbs.common.profiel
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties
 import jakarta.ws.rs.GET
@@ -12,12 +12,12 @@ import org.eclipse.microprofile.rest.client.inject.RegisterRestClient
 
 /**
  * REST-client naar de MOZA Profiel Service (zie github.com/MinBZK/moza-profiel-service).
- * We halen het volledige profiel van de ontvanger op en bepalen client-side of er een
- * voorkeur "OntvangViaBerichtenbox" bestaat met een scope die naar de afzender wijst.
+ * Gedeelde dependency voor berichtenmagazijn (validatie-flow) en berichtensessiecache
+ * (MagazijnResolver-flow); één Profiel-upstream betekent één client-definitie.
  *
  * De profiel-service zet identificatie in pad-parameters — afwijkend van onze interne
  * PII-richtlijn ("BSN nooit in URL"). We zijn hier gebonden aan het externe contract;
- * binnen de berichtenmagazijn-service zelf gaat BSN niet in URL.
+ * binnen onze services zelf gaat BSN niet in URL.
  *
  * DTO's zijn een minimale subset van de upstream-schema's; `@JsonIgnoreProperties`
  * negeert velden die we niet gebruiken (createdAt, lastUpdated, contactgegevens, etc.)
