+++ services/berichtenmagazijn/src/test/kotlin/nl/rijksoverheid/moz/fbs/berichtenmagazijn/aanlever/CircuitBreakerSkipOnTest.kt
@@ -10,8 +10,8 @@ import jakarta.inject.Inject
 import nl.rijksoverheid.moz.fbs.berichtenmagazijn.opslag.Bericht
 import nl.rijksoverheid.moz.fbs.berichtenmagazijn.opslag.BerichtRepository
 import nl.rijksoverheid.moz.fbs.berichtenmagazijn.validatie.MockProfielServiceClient
-import nl.rijksoverheid.moz.fbs.berichtenmagazijn.validatie.PartijResponse
-import nl.rijksoverheid.moz.fbs.berichtenmagazijn.validatie.ProfielServiceClient
+import nl.rijksoverheid.moz.fbs.common.profiel.PartijResponse
+import nl.rijksoverheid.moz.fbs.common.profiel.ProfielServiceClient
 import nl.rijksoverheid.moz.fbs.common.exception.DomainValidationException
 import org.eclipse.microprofile.rest.client.inject.RestClient
 import org.hibernate.exception.ConstraintViolationException as HibernateConstraintViolationException
